data/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pausemap"
version = "0.1.0"
authors = [{ name = "You" }]
description = "Data processing for pause.map"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "google-cloud-bigquery>=3.11.0",
    "tqdm>=4.65.0",
    "polars>=0.20.0",
    "pyarrow>=14.0.1",
    "fastparquet>=2023.10.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.1",
    "pytest-cov>=4.1.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=pausemap"